+++ DataLayer/LargeDatabase/SharedEntity.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DataLayer.LargeDatabase
+{
+    public class SharedEntity
+    {
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+    }
+}