+++ Test/UnitTests/TestDataLayer/TestPostgreSqlHelpers.cs
@@ -10,6 +10,7 @@ using Npgsql;
 using Test.Helpers;
 using TestSupport.Attributes;
 using TestSupport.EfHelpers;
+using TestSupport.Helpers;
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Extensions.AssertExtensions;
@@ -63,18 +64,79 @@ namespace Test.UnitTests.TestDataLayer
             using var context = new BookContext(options);
 
             context.Database.EnsureCreated();
+            context.SeedDatabaseFourBooks();
 
+            //ATTEMPT
             using (new TimeThings(_output, "Time to EnsureDeleted and EnsureCreated"))
             {
                 context.Database.EnsureDeleted();
                 context.Database.EnsureCreated();
             }
 
-            //ATTEMPT
+            //VERIFY
+            context.Books.Count().ShouldEqual(0);
+        }
+
+        [Fact]
+        public void TestEnsureCleanExistingDatabaseOk()
+        {
+            //SETUP
+            var options = this.CreatePostgreSqlUniqueDatabaseOptions<BookContext>();
+            using var context = new BookContext(options);
+
+            context.Database.EnsureCreated(); 
             context.SeedDatabaseFourBooks();
 
+            //ATTEMPT
+            using (new TimeThings(_output, "Time to EnsureClean"))
+            {
+                context.Database.EnsureClean();
+            }
+
             //VERIFY
-            context.Books.Count().ShouldEqual(4);
+            context.Books.Count().ShouldEqual(0);
+        }
+
+
+        //This proves that PostgreSQL EnsureClean doesn't delete the default named migration history table
+        //but it does delete migration history tables 
+        //To check that it doesn't delete the default named migration history table remove the
+        //MigrationsHistoryTable settings. If you run it twice it will NOT update the database schema because its already applied
+        [RunnableInDebugOnly]
+        public void TestEnsureCleanApplyMigrationOk()
+        {
+            //SETUP
+            var connectionString = this.GetUniquePostgreSqlConnectionString();
+            var optionsBuilder = new DbContextOptionsBuilder<BookContext>();
+            optionsBuilder.UseNpgsql(connectionString, dbOptions =>
+                dbOptions.MigrationsHistoryTable("BookMigrationHistoryName"));
+            using var context = new BookContext(optionsBuilder.Options);
+
+            //ATTEMPT      
+            context.Database.EnsureClean(false);
+            context.Database.Migrate();
+
+            //VERIFY
+            context.Books.Count().ShouldEqual(0);
+        }
+
+        [Fact]
+        public void TestEnsureCleanNoExistingDatabaseOk()
+        {
+            //SETUP
+            var options = this.CreatePostgreSqlUniqueDatabaseOptions<BookContext>();
+            using var context = new BookContext(options);
+
+            context.Database.EnsureDeleted(); 
+
+            //ATTEMPT
+            using (new TimeThings(_output, "Time to EnsureClean"))
+            {
+                context.Database.EnsureClean();
+            }
+
+            //VERIFY
+            context.Books.Count().ShouldEqual(0);
         }
 
         [Fact]
@@ -90,13 +152,14 @@ namespace Test.UnitTests.TestDataLayer
             var options = this.CreatePostgreSqlUniqueClassOptionsWithLogTo<BookContext>(log => _output.WriteLine(log), logOptions);
             using (var context = new BookContext(options))
             {
+                context.Database.EnsureDeleted();
                 context.Database.EnsureCreated();
                 context.SeedDatabaseFourBooks();
             }
             using (var context = new BookContext(options))
             {
                 //ATTEMPT
-                logOptions.ShowLog = true;
+                //logOptions.ShowLog = true;
                 using (new TimeThings(_output, "Time to empty database"))
                 {
                     await context.EnsureCreatedAndEmptyPostgreSqlAsync();